Port semantix to Python 3
After some consideration we have decided that Semantix should be written
using Python 3.  Being a from-scratch project it makes little sense to
care about backward compatibility.  And we do not use a lot of external
libs yet.  On the other hand, string and IO cleanups of Python 3 will
prevent a lot of headache.