Convert tests to py.test infrastructure
Standard unittest has clearly failed to satisfy our requirements for a
robust testsuite infrastructure.  Even with our limited number of
tests we had to invent hacks to extend and customize test discovery and
selection.  Moreover, setting up any kind of sane boilerplate for a
group of tests is a huge pain with unittest with its disconnected
setUp()/tearDown() methods.

Our next contender is py.test.  It has powerful built-in discovery
mechanisms, clever factory-based support for environment setup
(funcargs) and it is possible to extend it nicely without the need to
modify any tests.

I was hoping that it will also help speed up test execution, but
performance wins are marginal.

This commit has a balance of -337 lines of code, which clearly shows
what level of unnecessary provisioning overhead unittest had.