lang.yaml: Add support for automatic schema-controlled object construction
This commit makes it possible to specify in a schema that a certain node
has to be constructed as a given object.

Adding 'object: full.class.Name' to a node declaration in a schema will
attach a special tag to a maching data node during validation, which in
turn will call the specified class' 'construct' classmethod during
construction phase.