Add support for imports in YAML documents and use it to import Caos modules
Add support for %IMPORT directive to YAML documents.  The syntax is as
follows:

%IMPORT <module> [AS <alias>] [, <module> [AS <alias>]]*

In general, the requested modules will be imported and placed into
document context before the importing document is constructed.

Use this functionality to support Caos module imports.  The loader
performs several passes over the data to fully resolve the names and
cycles.