caos: Rework pgsql backend meta synchronization in terms of delta application
Instead of reinterpreting delta and creating a standalone command tree,
'adapt' the original delta tree and overload apply() method to produce
pgsql commands as a by-product of delta application.  This is a much
better approach, because instead of having two discrete points of
metadata state ('before' delta and 'after'), pgsql commands are now
generated on the fly in context of in-between deltas.

Other changes:
 * include proper link property support;
 * conversion of Prototype objects and delta Commands to Structs
 * conversion of backend-specific overrides of prototype objects and
   delta commands to the Adapter infrastructure