caos: Rework schema module processing
Current schema loading process uses the "common pot" approach, whereby
the schema and all its dependencies are added recursively to a common
stage and are processed simultaneously.  While this approach has certain
advantages (such as support for circular imports), it violates the
semantics of Python module loading and requires the schema to be loaded
via a special process instead of a simple import.  Furthermore, the
loaded module does not contain actual schema classes, but a raw
prototype schema, which is not at all useful for the application.

So, a rework of the schema module loading:

  * module is now a first-class schema prototypes and serves as
    a container for prototypes defined in said module;

  * prototype schema no longer holds object prototypes directly,
    holds a list of prototype modules instead, although common
    operations, such as add(), delete() and get() work as expected,
    provided that the names of objects in arguments can be resolved to a
    proper fully-qualified name;

  * there is now a default global prototype schema, which is populated
    by importing schema modules through an import statement or
    importlib.import_module; support for custom target schema through
    import context has been dropped for now;

  * root schema object classes are now accessible through schema module
    attributes, i.e Caos classes can be imported and used just like any
    regular Python class;

  * hacky schema cache implementation has been dropped: a proper
    multi-level cache system will need to be implemented;

  * it is no longer possible to override the name of the schema module
    through the "module" attribute.