import system: Adapt for importlib changes in python3.3
Python 3.3 includes significantly reorganized importlib (which is used
for __import__ by default now).  The Finder implementation is
considerably simpler now, as we simply pass a sequence of loaders and
file extensions to standard FileFinder and let it do the rest.  However,
this effectively disabled the extension conflict check that was
enforcing the rule that there can be not more than one module with a
given name across all languages.