yaml: Implement support for dynamic schema composition
It is now possible to use YAML tags to specify an arbitrary schema
for portions of YAML documents.  The parent document schema does not
need to know about those schemas beforehand.

Example:

   %schema schema_a
   %import schema_b, schema_c
   ---

   !schema_b TypeFromSchemaB:
       data

   !schema_c TypeFromSchemaC:
       data

   TypeFromSchemaA:
       data

In addition, the new NamespaceModuleSchema allows for automatic
population of module namespace with top-level document mapping
and obviates the need to use ProtoSchema.