utils.lang: Create the infrastructure for runtime derivatives
Formally introducing the concept of *language runtime*.

A language runtime is an environment in which a module is actually
executed as a useful unit.  Runtime is a property of a language,
and, correspondingly, a property of a module.

A module is said to be compatible with the runtime either if the
module's language runtime is a non-strict superclass, or if the
module has *derivatives* that are compatible with the given runtime.
It is considered to be an error to import a module that is not
compatible with importing module's runtime.

A runtime derivative module is a module that is adapted from its
source module to be compatible with a given runtime.

The mechanism of derivatives existed previously as a custom system
in JavaScript Loader that relied heavily on Resource infrastructure.
The JavaScript Loader has now been converted to use the new system.

The resource publisher has been rewritten to understand the concept of
runtimes and derivatives.  The build process is now a two-stage
operation.  First, the publisher iterates through the tree of packages
starting from the list specified in Node.packages and searches for
modules that have defined runtimes and are tagged.  Upon discovery of
such a module, it is added to the list and the search terminates for
that module's branch.  The result is a so-called list of
*runtime roots* grouped by module tags.

The list is then iterated for each tag recursively to satisfy all
imports and dispatch modules to correct resource buckets.