utils.lang: Enforce import runtime consistency
The import system will now raise an error if an attempt is made to
import a module that is incompatible with the runtime(s) of the
importing module.  Two modules are considered compatible if each of
importing module's runtimes is a relative of at least one of imported
module's runtimes through a non-abstract ancestor, including self.

Perform some necessary code movements to conform to the new rule.

In passing, give (S)CSS modules a default runtime so they can work as
runtime roots.