metamagic.test: Implement new test-system
New test system is a replacement for pytest.  The reasons to develop a
new test framework from scratch were:

 - complexity of pytest; lack of proper documentation of its internals
   and architecture;

 - broken support of python 3.3, and rather complicated process of
   fixing it and getting patches accepted;

 - need of new functionality -- benchmarks sub-system;

 - future extensibility & multi-language (pythonscript, javascript)
   reuse.

To run tests use the old '$ mm test' command.  It provides the same
interface to select/deselect tests, '--keep-going' and '--verbose'
options.

'metamagic.test' package contains 'skipif' and 'xfail' decorators, and a
'skip' method, which is used to skip a test during its execution.

To run benchmarks use '$ mm benchmark' command, which has similar
arguments to '$ mm test'.  'metamagic.test.benchmark' module exposes two
decorators:

 - 'throughput' to measure ops/second;

 - 'timing' to measure execution time.

The use of the above benchmark decorators is required.  The decorated
method should be named starting with 'benchmark_', and should return a
callable, which later will be benchmarked:

  @benchmark.throughput
  def benchmark_api_request():
      def call_remote_server():
          [...]
      return call_remote_server

The last thing to mention is function arguments.  We now use function
arguments instead of pytest's funcargs.  While the idea is roughly the
same, the usage is different:

  def test_db_connection(session:SessionArg):
      [...]

Where SessionArg is a class with two methods: 'setup' and 'teardown'.
Function arguments have different scope and are cached by the argument
name in their scope.

More on the new test system in the upcoming documentation ;)