utils.lang.import_.loader: Update sys.modules as soon as module is (re-)wrapped
YAML language allows its schemas to redefine the module class, so the
module object is recreated and repopulated relatively late in import
process.  Make sure sys.modules is updated as soon as this happens,
otherwise side-effects of the final part of import may go into the
original module object and be lost.  This particular defect was revealed
by the behaviour of LazyImportsModule, where imports are executed
_after_ the module is executed, so module dictionary modifications were
performed on an original module object and were discarded, leading to
obscure import errors.