boostrap.yaml.validator: Rewrite choice checker to avoid node copying
In order to be able discard any mutations caused by the failed choice
check tree, the choice checker is currently performing a deep copy of
a node, before trying a choice option on it.  Unfortunately, such an
approach makes use of YAML references in documents almost impossible, as
there is no way to prevent repeat validation (and potentially invalid
mutation) of the same node.

Fix this by rewriting the choice checker to use a two-step process.
First, as previously, a regular check attempt is ran on a tree copy.
The successful choice options are recorded in a choice profile tree,
which is then replayed on an original node.  This makes it possible to
keep track of checked nodes properly and avoid repeat mutation.